// nodes/Imagekit/resources/_generated/fileVersion.ts
const showOnlyForFileVersion = {
	resource: ['fileVersion'],
};

const showOnlyForFileVersionList = {
	operation: ['list'],
	resource: ['fileVersion'],
};

const showOnlyForFileVersionGet = {
	operation: ['get'],
	resource: ['fileVersion'],
};

const showOnlyForFileVersionDelete = {
	operation: ['delete'],
	resource: ['fileVersion'],
};

const showOnlyForFileVersionRestore = {
	operation: ['restore'],
	resource: ['fileVersion'],
};

export const fileVersionDescription: INodeProperties[] = [
	{
		displayName: 'Operation',
		name: 'operation',
		type: 'options',
		noDataExpression: true,
		displayOptions: { show: showOnlyForFileVersion },
		options: [
			{
				name: 'Delete',
				value: 'delete',
				action: 'Delete a file version',
				description: 'Delete a specific version of a file',
			},
			{
				name: 'Get',
				value: 'get',
				action: 'Get a file version',
				description: 'Get a specific version of a file',
			},
			{
				name: 'List',
				value: 'list',
				action: 'List file versions',
				description: 'List all versions of a file',
			},
			{
				name: 'Restore',
				value: 'restore',
				action: 'Restore a file version',
				description: 'Restore a specific version of a file',
			},
		],
		default: 'list',
	},
	// List
	{
		displayName: 'File ID',
		name: 'fileId',
		type: 'string',
		required: true,
		default: '',
		description: 'The unique ID of the file',
		displayOptions: { show: showOnlyForFileVersionList },
	},
	// Get
	{
		displayName: 'File ID',
		name: 'fileId',
		type: 'string',
		required: true,
		default: '',
		description: 'The unique ID of the file',
		displayOptions: { show: showOnlyForFileVersionGet },
	},
	{
		displayName: 'Version ID',
		name: 'versionId',
		type: 'string',
		required: true,
		default: '',
		description: 'The version ID to retrieve',
		displayOptions: { show: showOnlyForFileVersionGet },
	},
	// Delete
	{
		displayName: 'File ID',
		name: 'fileId',
		type: 'string',
		required: true,
		default: '',
		description: 'The unique ID of the file',
		displayOptions: { show: showOnlyForFileVersionDelete },
	},
	{
		displayName: 'Version ID',
		name: 'versionId',
		type: 'string',
		required: true,
		default: '',
		description: 'The version ID to delete',
		displayOptions: { show: showOnlyForFileVersionDelete },
	},
	// Restore
	{
		displayName: 'File ID',
		name: 'fileId',
		type: 'string',
		required: true,
		default: '',
		description: 'The unique ID of the file',
		displayOptions: { show: showOnlyForFileVersionRestore },
	},
	{
		displayName: 'Version ID',
		name: 'versionId',
		type: 'string',
		required: true,
		default: '',
		description: 'The version ID to restore',
		displayOptions: { show: showOnlyForFileVersionRestore },
	},
];