// nodes/Imagekit/resources/_generated/accountOrigin.ts
const showOnlyForAccountOrigin = {
	resource: ['accountOrigin'],
};

const showOnlyForAccountOriginCreate = {
	operation: ['create'],
	resource: ['accountOrigin'],
};

const showOnlyForAccountOriginGet = {
	operation: ['get'],
	resource: ['accountOrigin'],
};

const showOnlyForAccountOriginUpdate = {
	operation: ['update'],
	resource: ['accountOrigin'],
};

const showOnlyForAccountOriginDelete = {
	operation: ['delete'],
	resource: ['accountOrigin'],
};

export const accountOriginDescription: INodeProperties[] = [
	{
		displayName: 'Operation',
		name: 'operation',
		type: 'options',
		noDataExpression: true,
		displayOptions: { show: showOnlyForAccountOrigin },
		options: [
			{
				name: 'Create',
				value: 'create',
				action: 'Create an origin',
				description: 'Create a new origin',
			},
			{
				name: 'Delete',
				value: 'delete',
				action: 'Delete an origin',
				description: 'Delete an origin',
			},
			{
				name: 'Get',
				value: 'get',
				action: 'Get an origin',
				description: 'Get details of a specific origin',
			},
			{
				name: 'List',
				value: 'list',
				action: 'List origins',
				description: 'List all origins',
			},
			{
				name: 'Update',
				value: 'update',
				action: 'Update an origin',
				description: 'Update an existing origin',
			},
		],
		default: 'list',
	},
	// Create
	{
		displayName: 'Origin Data (JSON)',
		name: 'originData',
		type: 'json',
		required: true,
		default: '{}',
		description: 'JSON object with origin configuration',
		displayOptions: { show: showOnlyForAccountOriginCreate },
	},
	// Get
	{
		displayName: 'Origin ID',
		name: 'originId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the origin to get',
		displayOptions: { show: showOnlyForAccountOriginGet },
	},
	// Update
	{
		displayName: 'Origin ID',
		name: 'originId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the origin to update',
		displayOptions: { show: showOnlyForAccountOriginUpdate },
	},
	{
		displayName: 'Origin Data (JSON)',
		name: 'originData',
		type: 'json',
		required: true,
		default: '{}',
		description: 'JSON object with origin configuration to update',
		displayOptions: { show: showOnlyForAccountOriginUpdate },
	},
	// Delete
	{
		displayName: 'Origin ID',
		name: 'originId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the origin to delete',
		displayOptions: { show: showOnlyForAccountOriginDelete },
	},
];