// nodes/Imagekit/resources/_generated/accountUrlEndpoint.ts
const showOnlyForAccountUrlEndpoint = {
	resource: ['accountUrlEndpoint'],
};

const showOnlyForAccountUrlEndpointCreate = {
	operation: ['create'],
	resource: ['accountUrlEndpoint'],
};

const showOnlyForAccountUrlEndpointGet = {
	operation: ['get'],
	resource: ['accountUrlEndpoint'],
};

const showOnlyForAccountUrlEndpointUpdate = {
	operation: ['update'],
	resource: ['accountUrlEndpoint'],
};

const showOnlyForAccountUrlEndpointDelete = {
	operation: ['delete'],
	resource: ['accountUrlEndpoint'],
};

export const accountUrlEndpointDescription: INodeProperties[] = [
	{
		displayName: 'Operation',
		name: 'operation',
		type: 'options',
		noDataExpression: true,
		displayOptions: { show: showOnlyForAccountUrlEndpoint },
		options: [
			{
				name: 'Create',
				value: 'create',
				action: 'Create a URL endpoint',
				description: 'Create a new URL endpoint',
			},
			{
				name: 'Delete',
				value: 'delete',
				action: 'Delete a URL endpoint',
				description: 'Delete a URL endpoint',
			},
			{
				name: 'Get',
				value: 'get',
				action: 'Get a URL endpoint',
				description: 'Get details of a specific URL endpoint',
			},
			{
				name: 'List',
				value: 'list',
				action: 'List URL endpoints',
				description: 'List all URL endpoints',
			},
			{
				name: 'Update',
				value: 'update',
				action: 'Update a URL endpoint',
				description: 'Update an existing URL endpoint',
			},
		],
		default: 'list',
	},
	// Create
	{
		displayName: 'Endpoint Data (JSON)',
		name: 'endpointData',
		type: 'json',
		required: true,
		default: '{}',
		description: 'JSON object with URL endpoint configuration',
		displayOptions: { show: showOnlyForAccountUrlEndpointCreate },
	},
	// Get
	{
		displayName: 'Endpoint ID',
		name: 'endpointId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the URL endpoint to get',
		displayOptions: { show: showOnlyForAccountUrlEndpointGet },
	},
	// Update
	{
		displayName: 'Endpoint ID',
		name: 'endpointId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the URL endpoint to update',
		displayOptions: { show: showOnlyForAccountUrlEndpointUpdate },
	},
	{
		displayName: 'Endpoint Data (JSON)',
		name: 'endpointData',
		type: 'json',
		required: true,
		default: '{}',
		description: 'JSON object with URL endpoint configuration to update',
		displayOptions: { show: showOnlyForAccountUrlEndpointUpdate },
	},
	// Delete
	{
		displayName: 'Endpoint ID',
		name: 'endpointId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the URL endpoint to delete',
		displayOptions: { show: showOnlyForAccountUrlEndpointDelete },
	},
];