// nodes/Imagekit/resources/_generated/savedExtension.ts
const showOnlyForSavedExtension = {
	resource: ['savedExtension'],
};

const showOnlyForSavedExtensionCreate = {
	operation: ['create'],
	resource: ['savedExtension'],
};

const showOnlyForSavedExtensionGet = {
	operation: ['get'],
	resource: ['savedExtension'],
};

const showOnlyForSavedExtensionUpdate = {
	operation: ['update'],
	resource: ['savedExtension'],
};

const showOnlyForSavedExtensionDelete = {
	operation: ['delete'],
	resource: ['savedExtension'],
};

export const savedExtensionDescription: INodeProperties[] = [
	{
		displayName: 'Operation',
		name: 'operation',
		type: 'options',
		noDataExpression: true,
		displayOptions: { show: showOnlyForSavedExtension },
		options: [
			{
				name: 'Create',
				value: 'create',
				action: 'Create a saved extension',
				description: 'Create a new saved extension',
			},
			{
				name: 'Delete',
				value: 'delete',
				action: 'Delete a saved extension',
				description: 'Delete a saved extension',
			},
			{
				name: 'Get',
				value: 'get',
				action: 'Get a saved extension',
				description: 'Get details of a specific saved extension',
			},
			{
				name: 'List',
				value: 'list',
				action: 'List saved extensions',
				description: 'List all saved extensions',
			},
			{
				name: 'Update',
				value: 'update',
				action: 'Update a saved extension',
				description: 'Update an existing saved extension',
			},
		],
		default: 'list',
	},
	// Create
	{
		displayName: 'Extension Data (JSON)',
		name: 'extensionData',
		type: 'json',
		required: true,
		default: '{}',
		description: 'JSON object with saved extension configuration',
		displayOptions: { show: showOnlyForSavedExtensionCreate },
	},
	// Get
	{
		displayName: 'Extension ID',
		name: 'extensionId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the saved extension to get',
		displayOptions: { show: showOnlyForSavedExtensionGet },
	},
	// Update
	{
		displayName: 'Extension ID',
		name: 'extensionId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the saved extension to update',
		displayOptions: { show: showOnlyForSavedExtensionUpdate },
	},
	{
		displayName: 'Extension Data (JSON)',
		name: 'extensionData',
		type: 'json',
		required: true,
		default: '{}',
		description: 'JSON object with saved extension configuration to update',
		displayOptions: { show: showOnlyForSavedExtensionUpdate },
	},
	// Delete
	{
		displayName: 'Extension ID',
		name: 'extensionId',
		type: 'string',
		required: true,
		default: '',
		description: 'The ID of the saved extension to delete',
		displayOptions: { show: showOnlyForSavedExtensionDelete },
	},
];